requirements.txt
annotated-types==0.7.0
anyio==4.5.2
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2025.11.12
distro==1.9.0
Django==4.2.26
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.9.1
openai==2.2.0
pillow==10.4.0
pkg_resources==0.0.0
pydantic==2.10.6
pydantic_core==2.27.2
PyPDF2==3.0.1
python-dotenv==1.0.1
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.13.2
